// src/app/api/model/pedido.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { Plato } from './plato';


export interface Pedido { 
    id?: string;
    plato?: Plato;
    user?: User;
    modifications?: Array<string>;
    estado?: Pedido.EstadoEnum;
    entrega?: string;
}
export namespace Pedido {
    export const EstadoEnum = {
        Preparing: 'PREPARING',
        Delivering: 'DELIVERING',
        Delivered: 'DELIVERED'
    } as const;
    export type EstadoEnum = typeof EstadoEnum[keyof typeof EstadoEnum];
}